+++ jinahub/indexers/searcher/compound/FaissPostgresSearcher/faisspsql.py
@@ -8,7 +8,7 @@ from jina import requests, DocumentArray, Executor
 
 from jina_commons import get_logger
 from jinahub.indexers.searcher.FaissSearcher.faiss_searcher import FaissSearcher
-from jinahub.indexers.storage.PostgreSQLStorage.postgres import PostgreSQLStorage
+from jinahub.indexers.storage.PostgreSQLStorage import PostgreSQLStorage
 
 
 class FaissPostgresSearcher(Executor):
