+++ jinahub/indexers/storage/PostgreSQLStorage/config.yml
@@ -3,8 +3,8 @@ with:
   {}
 metas:
   py_modules:
+    - __init__.py
     - postgreshandler.py
-    - postgres.py
   workspace: workspace/
   name: psql
