+++ jinahub/indexers/storage/PostgreSQLStorage/tests/test_postgres_dbms.py
@@ -7,7 +7,7 @@ from jina import Document, DocumentArray, Flow
 from jina.logging.profile import TimeContext
 from jina_commons.indexers.dump import import_vectors, import_metas
 
-from ..postgres import PostgreSQLStorage
+from .. import PostgreSQLStorage
 from ..postgreshandler import doc_without_embedding
