+++ tests/integration/psql_dump_reload/test_dump_psql.py
@@ -38,7 +38,7 @@ from jinahub.indexers.storage.PostgreSQLStorage.postgreshandler import (
 from jinahub.indexers.searcher.compound.FaissPostgresSearcher import (
     FaissPostgresSearcher,
 )
-from jinahub.indexers.storage.PostgreSQLStorage.postgres import PostgreSQLStorage
+from jinahub.indexers.storage.PostgreSQLStorage import PostgreSQLStorage
 
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 compose_yml = os.path.join(cur_dir, 'docker-compose.yml')
